distribute functions into modules to a package structure
